+++ packages/f-content-cards/src/components/_tests/ContentCards.test.js
@@ -189,7 +189,7 @@ describe('ContentCards', () => {
         const arrange = async () => {
             const PromotionCard = Vue.extend({
                 template: '<div data-promotion-card="true"></div>',
-                inject: ['emitCardClick', 'emitCardView', 'emitVoucherCodeClicked']
+                inject: ['emitCardClick', 'emitCardView', 'emitVoucherCodeClick']
             });
             const cardTypes = ['Promotion_Card_1', 'Promotion_Card_2', 'Post_Order_Card_1'];
             const appboy = createAppboyInstance(cardTypes);
@@ -205,7 +205,7 @@ describe('ContentCards', () => {
             });
             const cardClickHandler = jest.spyOn(instance.vm._provided, 'emitCardClick');
             const cardViewHandler = jest.spyOn(instance.vm._provided, 'emitCardView');
-            const voucherCodeClickedHandler = jest.spyOn(instance.vm._provided, 'emitVoucherCodeClicked');
+            const voucherCodeClickHandler = jest.spyOn(instance.vm._provided, 'emitVoucherCodeClick');
             instance.vm.contentCards(appboy);
             await instance.vm.$nextTick();
 
@@ -213,7 +213,7 @@ describe('ContentCards', () => {
                 instance,
                 cardClickHandler,
                 cardViewHandler,
-                voucherCodeClickedHandler
+                voucherCodeClickHandler
             };
         };
 
@@ -303,26 +303,26 @@ describe('ContentCards', () => {
         });
 
         it('should provide a voucher code copied handler', async () => {
-            const { instance, voucherCodeClickedHandler } = await arrange();
+            const { instance, voucherCodeClickHandler } = await arrange();
 
             // Act
-            instance.find('[data-promotion-card="true"]').vm.emitVoucherCodeClicked(url);
+            instance.find('[data-promotion-card="true"]').vm.emitVoucherCodeClick(url);
 
             // Assert
-            expect(voucherCodeClickedHandler).toHaveBeenCalledWith(url);
+            expect(voucherCodeClickHandler).toHaveBeenCalledWith(url);
         });
 
         describe('the voucher code copied handler', () => {
-            it('should emit a voucherCodeClicked event', async () => {
+            it('should emit a voucherCodeClick event', async () => {
                 const { instance } = await arrange();
 
                 // Act
-                instance.find('[data-promotion-card="true"]').vm.emitVoucherCodeClicked(url);
+                instance.find('[data-promotion-card="true"]').vm.emitVoucherCodeClick(url);
 
                 // Assert
-                expect(instance.emitted().voucherCodeClicked).toBeTruthy();
-                expect(instance.emitted().voucherCodeClicked.length).toBe(1);
-                expect(instance.emitted().voucherCodeClicked[0]).toEqual([{
+                expect(instance.emitted().voucherCodeClick).toBeTruthy();
+                expect(instance.emitted().voucherCodeClick.length).toBe(1);
+                expect(instance.emitted().voucherCodeClick[0]).toEqual([{
                     url
                 }]);
             });