+++ packages/f-content-cards/src/tenants/da-DK.js
@@ -1,0 +1,4 @@
+export default {
+    locale: 'da-DK',
+    copyCodeLabel: 'Kopier kode'
+};