+++ packages/f-content-cards/CHANGELOG.md
@@ -4,9 +4,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-v1.5.0
+v1.6.0
 ------------------------------
-*July 28, 2020*
+*July 31, 2020*
 
 ### Added
 
@@ -15,6 +15,18 @@ v1.5.0
     - `@get-card-count`
     - `@has-loaded`
     - `@on-error`
+
+
+v1.5.0
+------------------------------
+*July 31, 2020*
+
+### Added
+- `VoucherCard` component from HomeWeb's Offer page
+- Associated stories for Voucher and Anniversary card types
+- I18n config for 'Copy Code' button
+- `locale` prop to enable above to take effect
+
 
 v1.4.0
 ------------------------------